[REFACTOR] EntireTrack과 Track 분리 및 props 정리
