+++ src/components/attendance/completed/component/componentType.ts
@@ -1,0 +1,8 @@
+import { MemberStackKor } from '@@types/request';
+
+export interface ITrackController {
+  title: MemberStackKor;
+  arrayLength: number;
+  titleColor: string;
+  color: string;
+}
